src/datamgr/metadata/local_requirements.txt
about_time==2.0.5
arpeggio==1.9.2
arrow~=0.15.5
attrdict~=2.0.1
attrs~=19.1.0
babel==2.8.0
backports.functools-lru-cache==1.6.1
boto3==1.11.15
boto==2.49.0
botocore==1.14.15
bz2file==0.98
cached-property==1.5.1
cattrs==1.0.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.1
diskcache==4.1.0
docutils==0.15.2
enum34~=1.1.6
flask==1.1.1
gensim==3.8.1
getmac~=0.8.2
gevent~=1.4.0
gipc~=1.0.1
greenlet==0.4.15 ; platform_python_implementation == 'CPython'
huey~=2.1.3
idna==2.8
importlib-metadata==1.5.0
inflect==3.0.2
invoke~=1.4.1
itsdangerous==1.1.0
jieba~=0.42.1
jinja2~=2.11.1
jmespath==0.9.4
jsonext~=0.4.2
kafka-python==2.0.0
kazoo~=2.6.1
lazy-import==0.2.2
markupsafe==1.1.1
more-itertools==5.0.0
networkx~=2.2
nltk==3.4.5
numpy~=1.16.1
pandas==0.24.2
parse~=1.14.0
portalocker==1.5.2
psutil~=5.6.7
py==1.8.1
pycrypto==2.6.1
pyenchant==2.0.0
pymysql~=0.9.3
python-dateutil~=2.8.1
pytz~=2019.3
redis==3.4.1
# redislite==5.0.165407
requests-toolbelt==0.9.1
requests~=2.22.0
retry~=0.9.2
s3transfer==0.3.3
scikit-learn~=0.20.4
scipy~=1.2.3
simplejson==3.17.0
singledispatch~=3.4.0.3
six~=1.14.0
smart-open==1.9.0
speaklater3~=1.4
sqlacodegen==2.1.0
sqlalchemy-utils==0.36.1
sqlalchemy~=1.3.13
tblib~=1.6.0
textx~=2.1.0
typing~=3.7.4.1
urllib3==1.25.8 ; python_version != '3.4'
uwsgi==2.0.18
walrus==0.8.0
werkzeug~=1.0.0
cachelib==0.1
wordsegment==1.3.1
zipp==1.1.0
kombu~=4.6.11
tinyrpc~=1.0.4

PyYAML~=5.4.1
attr~=0.3.1
Click~=7.0
setuptools~=44.1.0